Fixed a verrrrrrry frustrating issue with new id convert wf
Apparently plink2 does not auto-sort the output bed, which is a problem.
Must use plink1.9 make-bed for extract_chrs to ensure bim is sorted in numerical chr sort order
Was causing the silent issue of the workflow 'succeeding' but the new bim wasn't aligned to the old bed
Meaning the genotypes were basically...meaningless. So frustrating.